fix: move Copy as Markdown button to session-turn timeline
The assistant's text parts are rendered through session-turn.tsx's
timeline (not AssistantMessageDisplay), so the copy button must live
in the turn-level timeline area, placed after all parts and diffs.
